Move computed_task_reported() to golem.task.server.helpers
